Settings.check_dependencies to verify dependencies
