+++ neo/Settings.py
@@ -9,12 +9,17 @@ reconfigure them by calling the `setup(..)` methods.
 import json
 import os
 import logging
+import pip
+
 from json.decoder import JSONDecodeError
+import logzero
 
 from neo import __version__
 from neocore.Cryptography import Helper
 
-import logzero
+from neorpc.Client import RPCClient
+from neorpc.Settings import settings as rpc_settings
+
 
 # Create am absolute references to the project root folder. Used for
 # specifying the various filenames.
@@ -30,6 +35,51 @@ FILENAME_SETTINGS_MAINNET = os.path.join(DIR_PROJECT_ROOT, 'protocol.mainnet.jso
 FILENAME_SETTINGS_TESTNET = os.path.join(DIR_PROJECT_ROOT, 'protocol.testnet.json')
 FILENAME_SETTINGS_PRIVNET = os.path.join(DIR_PROJECT_ROOT, 'protocol.privnet.json')
 FILENAME_SETTINGS_COZNET = os.path.join(DIR_PROJECT_ROOT, 'protocol.coz.json')
+
+
+class PrivnetConnectionError(Exception):
+    pass
+
+
+class DependencyError(Exception):
+    pass
+
+
+def check_privatenet():
+    """ Check if privatenet is running, and if container is same as the chain file """
+    rpc_settings.setup(settings.RPC_LIST)
+    client = RPCClient()
+    version = client.get_version()
+    if not version:
+        raise PrivnetConnectionError("Error: private network container doesn't seem to be running, or RPC is not enabled.")
+
+    print("Privatenet useragent '%s', nonce: %s" % (version["useragent"], version["nonce"]))
+
+    # Now check if nonce is the same as in the chain path
+    nonce_container = str(version["nonce"])
+    neopy_chain_meta_filename = os.path.join(settings.LEVELDB_PATH, ".privnet-nonce")
+    if os.path.isfile(neopy_chain_meta_filename):
+        nonce_chain = open(neopy_chain_meta_filename, "r").read()
+        if nonce_chain != nonce_container:
+            raise PrivnetConnectionError(
+                "Chain database in Chains/privnet is for a different private network than the current container. "
+                "Consider deleting the Chain directory with 'rm -rf Chains/privnet*'."
+            )
+    else:
+        with open(neopy_chain_meta_filename, "w") as f:
+            f.write(nonce_container)
+
+
+def check_depdendencies():
+    installed_packages = pip.get_installed_distributions()
+    installed_packages_list = sorted(["%s==%s" % (i.key, i.version) for i in installed_packages])
+
+    # Now check if each required package is actually available
+    deps_filename = os.path.join(DIR_PROJECT_ROOT, "requirements.txt")
+    deps_list = open(deps_filename, "r").read().split()
+    for dep in deps_list:
+        if not dep.lower() in installed_packages_list:
+            raise DependencyError("Required dependency %s is not installed. Please run 'pip install -e .'." % dep)
 
 
 class SettingsHolder:
@@ -150,6 +200,7 @@ class SettingsHolder:
     def setup_privnet(self):
         """ Load settings from the privnet JSON config file """
         self.setup(FILENAME_SETTINGS_PRIVNET)
+        check_privatenet()
 
     def setup_coznet(self):
         """ Load settings from the coznet JSON config file """
@@ -191,3 +242,6 @@ settings.setup_testnet()
 
 # By default, set loglevel to INFO. DEBUG just print a lot of internal debug statements
 settings.set_loglevel(logging.INFO)
+
+# Check if currently installed dependencies match the requirements
+check_depdendencies()
