+++ prompt.py
@@ -40,13 +40,10 @@ from neo.Prompt.Commands.Wallet import DeleteAddress, ImportWatchAddr, ImportTok
     ShowUnspentCoins
 from neo.Prompt.Utils import get_arg
 from neo.Prompt.InputParser import InputParser
-from neo.Settings import settings, DIR_PROJECT_ROOT
+from neo.Settings import settings, DIR_PROJECT_ROOT, PrivnetConnectionError, DependencyError
 from neo.UserPreferences import preferences
 from neocore.KeyPair import KeyPair
 from neocore.UInt256 import UInt256
-
-from neorpc.Client import RPCClient
-from neorpc.Settings import settings as rpc_settings
 
 # Logfile settings & setup
 LOGFILE_FN = os.path.join(DIR_PROJECT_ROOT, 'prompt.log')
@@ -56,10 +53,6 @@ settings.set_logfile(LOGFILE_FN, LOGFILE_MAX_BYTES, LOGFILE_BACKUP_COUNT)
 
 # Prompt history filename
 FILENAME_PROMPT_HISTORY = os.path.join(DIR_PROJECT_ROOT, '.prompt.py.history')
-
-
-class PrivnetConnectionError(Exception):
-    pass
 
 
 class PromptInterface(object):
@@ -942,31 +935,6 @@ class PromptInterface(object):
                 traceback.print_exc()
 
 
-def check_privatenet():
-    """ Check if privatenet is running, and if container is same as the chain file """
-    rpc_settings.setup(settings.RPC_LIST)
-    client = RPCClient()
-    version = client.get_version()
-    if not version:
-        raise PrivnetConnectionError("Error: private network container doesn't seem to be running, or RPC is not enabled.")
-
-    print("Privatenet useragent '%s', nonce: %s" % (version["useragent"], version["nonce"]))
-
-    # Now check if nonce is the same as in the chain path
-    nonce_container = str(version["nonce"])
-    neopy_chain_meta_filename = os.path.join(settings.LEVELDB_PATH, ".privnet-nonce")
-    if os.path.isfile(neopy_chain_meta_filename):
-        nonce_chain = open(neopy_chain_meta_filename, "r").read()
-        if nonce_chain != nonce_container:
-            raise PrivnetConnectionError(
-                "Chain database in Chains/privnet is for a different private network than the current container. "
-                "Consider deleting the Chain directory with 'rm -rf Chains/privnet*'."
-            )
-    else:
-        with open(neopy_chain_meta_filename, "w") as f:
-            f.write(nonce_container)
-
-
 def main():
     parser = argparse.ArgumentParser()
 
@@ -1001,7 +969,11 @@ def main():
     elif args.mainnet:
         settings.setup_mainnet()
     elif args.privnet:
-        settings.setup_privnet()
+        try:
+            settings.setup_privnet()
+        except PrivnetConnectionError as e:
+            logger.error(str(e))
+            return
     elif args.coznet:
         settings.setup_coznet()
 
@@ -1010,14 +982,6 @@ def main():
 
     if args.verbose:
         settings.set_log_smart_contract_events(True)
-
-    # If privnet, check if correct chain
-    if args.privnet:
-        try:
-            check_privatenet()
-        except PrivnetConnectionError as e:
-            logger.error(str(e))
-            return
 
     # Instantiate the blockchain and subscribe to notifications
     blockchain = LevelDBBlockchain(settings.LEVELDB_PATH)
